apps/go_django/requirements.txt
alabaster==0.7.6
argcomplete==1.0.0
Babel==2.3.4
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.8.3
bokeh==0.10.0
boto==2.38.0
Bottleneck==1.0.0
cffi==1.2.1
clyent==1.2.0
colorama==0.3.3
comtypes==1.1.2
configobj==5.0.6
coverage==4.1
cryptography==1.0.1
cycler==0.9.0
Cython==0.23.4
cytoolz==0.7.4
datashape==0.4.7
decorator==4.0.6
Django==1.10
django-debug-toolbar==1.4
django-nested-admin==3.0.10
django-phonenumber-field==1.1.0
django-simple-history==1.8.1
django-toolbelt==0.0.1
django-webpack-loader==0.3.0
djangorestframework==3.3.2
docutils==0.12
fastcache==1.0.2
Flask==0.10.1
greenlet==0.4.9
gunicorn==19.6.0
httpie==0.9.4
idna==2.0
itsdangerous==0.24
jedi==0.9.0
Jinja2==2.8
jsonschema==2.4.0
lxml==3.6.0
markdown2==2.3.1
MarkupSafe==0.23
mistune==0.7.1
multipledispatch==0.4.8
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.10
nose==1.3.7
notebook==4.1.0
odo==0.3.4
openpyxl==2.2.6
patsy==0.4.0
pep8==1.6.2
phonenumberslite==7.4.5
pickleshare==0.5
Pillow==3.0.0
ply==3.8
psutil==3.3.0
psycopg2==2.6.1
py==1.4.30
pyasn1==0.1.9
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
pyflakes==1.0.0
Pygments==2.0.2
pykml==0.1.0
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyreadline==2.1
pytest==2.8.1
python-dateutil==2.4.2
pytz==2016.6.1
PyYAML==3.11
pyzmq==15.2.0
qtconsole==4.1.1
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.0
sockjs-tornado==1.0.1
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
spyder==2.3.8
SQLAlchemy==1.0.9
sqlparse==0.2.0
sympy==0.7.6.1
toolz==0.7.4
tornado==4.3
traitlets==4.1.0
ujson==1.33
unicodecsv==0.14.1
virtualenv==13.0.1
Werkzeug==0.11.2
whitenoise==3.2
xlrd==0.9.4
XlsxWriter==0.8.4
xlwt==1.0.0
